Toutes les fonctions test le nombdre de variable passées en paramètre avant de s'executer.

+++ SCRIPT/fonctionnel/del_blog.sh
@@ -5,4 +5,9 @@
 name=$1
 
 #Script
-/bin/rm -R /var/sftp/$name/www/*
+
+if (($#=="1"));then
+	/bin/rm -R /var/sftp/$name/www/*
+else
+	echo " del_blog: Nombre de paramètres invalide "
+fi
