+++ SCRIPT/fonctionnel/del_userMail.sh
@@ -6,14 +6,18 @@
 name=$1
 
 #Script
-#Retrait sur postfix
-/bin/sed -i "/${name}/d" /etc/postfix/mailboxmap
-/usr/sbin/postmap /etc/postfix/mailboxmap
-/usr/bin/service postfix reload
+if (($#=="1"));then
+	#Retrait sur postfix
+	/bin/sed -i "/${name}/d" /etc/postfix/mailboxmap
+	/usr/sbin/postmap /etc/postfix/mailboxmap
+	/usr/bin/service postfix reload
 
-#Retrait sur courier-IMAP
-/bin/sed -i "/${name}/d" /etc/courier/userdb
-/usr/sbin/makeuserdb
+	#Retrait sur courier-IMAP
+	/bin/sed -i "/${name}/d" /etc/courier/userdb
+	/usr/sbin/makeuserdb
 
-#Retrait du dossier mail utilisateur
-/bin/rm -r /var/mail/$name
+	#Retrait du dossier mail utilisateur
+	/bin/rm -r /var/mail/$name
+else
+	echo " del_userMail: Nombre de paramètre invalide "
+fi
