+++ SCRIPT/fonctionnel/del_userUnix.sh
@@ -6,8 +6,13 @@
 name=$1
 
 #Script
-#Suppression de l'utilisateur
-/usr/sbin/userdel --force $name 
 
-#Suppression de son répertoire
-/bin/rm -rf /var/sftp/home/$name
+if (($#=="1"));then
+	#Suppression de l'utilisateur
+	/usr/sbin/userdel --force $name 
+
+	#Suppression de son répertoire
+	/bin/rm -rf /var/sftp/home/$name
+else
+	echo " del_userUnix: Nombre de paramètre invalide "
+fi
