+++ SCRIPT/fonctionnel/del_vhost.sh
@@ -8,7 +8,11 @@
 name=$1
 
 #Script
-/bin/rm /etc/apache2/sites-available/$name.conf		# Deleting the .conf file
-/bin/rm -R /var/sftp/$name							# Deleting the directory
 
-/usr/bin/service apache2 reload							# Reloading the Apache service
+if (($#=="1"));then
+	/bin/rm /etc/apache2/sites-available/$name.conf		# Deleting the .conf file
+	/bin/rm -R /var/sftp/$name				# Deleting the directory
+	/usr/bin/service apache2 reload				# Reloading the Apache service
+else
+	echo " del_vhost: Nombre de paramètres invalide "
+fi
